+++ index.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
-const measurementsRoute = require("./route/measurements");
-const deviceRoute = require("./route/device");
+const measurementsRoute = require("./routes/measurements");
+const devicesRoute = require("./routes/devices");
 
 const app = express();
 const port = 3000;
@@ -9,7 +9,7 @@ const port = 3000;
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 
-app.use("/devices", deviceRoute);
+app.use("/devices", devicesRoute);
 app.use("/measurements", measurementsRoute);
 
 app.listen(port);
